commit with somnath

+++ api/po.py
@@ -920,9 +920,9 @@ async def approvepo(id:approvePO):
                         # qty = int(result_pur['msg'][0]['approved_ord_qty']) + int(item['quantity'])
                         # print(qty)
 
-                        pur_qty = float(result_pur['msg'][0]['ordered_qty']) 
-                        appr_qty =  float(result_pur['msg'][0]['approved_ord_qty']) 
-                        sum_qty = float(item['quantity'])
+                        pur_qty = Decimal(result_pur['msg'][0]['ordered_qty']) 
+                        appr_qty =  Decimal(result_pur['msg'][0]['approved_ord_qty']) 
+                        sum_qty = Decimal(item['quantity'])
 
                         if (pur_qty- appr_qty)<=sum_qty and sum_qty>0:
                                         approved_ord_qty = pur_qty - appr_qty
@@ -1292,8 +1292,8 @@ async def addfreshpo(data:PoModel):
                 result1 = await db_Insert(table_name1, fields1, values1, whr1, flag1)
                 item_save=1 if result1['suc']>0 else 0
 
-                sum_qty = int(c.qty)
-                ordered_qty = 0
+                sum_qty = Decimal(c.qty)
+                ordered_qty = 0.000
                 # if data.sl_no==0:
                 # for pur_qty in result_pur['msg']:
